+++ src/core/client-response.model.ts
@@ -1,0 +1,5 @@
+import { DTOModels } from './dto-name.model';
+
+export type ClientResponse<T extends DTOModels> = {
+  data: { [key: string]: T }[]
+}