+++ src/core/dto-name.model.ts
@@ -14,10 +14,7 @@ export type DTOName = 'Activity'
   | 'CompanyInfo'
   | 'CompanySettings'
   | 'Contact'
-  | 'CrowdAssignment'
-  | 'CrowdBusinessPartner'
   | 'CrowdExecutionRecord'
-  | 'CrowdPerson'
   | 'Currency'
   | 'CustomRule'
   | 'EmployeeBranch'
@@ -81,7 +78,7 @@ export type DTOName = 'Activity'
   | 'WorkTimePattern'
   | 'WorkTimeTask'
 
-export type DTOModels = {
+export type DTOModels = Partial<{
   id: undefined | null | string;
   lastChanged: undefined | null | number;
-} & Object;
+} & Object>;