+++ src/index.ts
@@ -1,2 +1,2 @@
-import { CoreAPIClient } from './CoreAPIClient';
+import { CoreAPIClient } from './core-api.client';
 export { CoreAPIClient };