+++ src/model/ClientResponse.ts
@@ -1,4 +1,0 @@
-export type ClientResponse<T> = {
-  data: { [key: string]: T }[]
-}
-