+++ src/test/data-api.spec.ts
@@ -1,6 +1,6 @@
 import assert = require('assert');
-import { integrationTestConfig } from './integrationTest.config';
-import { CoreAPIClient } from '../CoreAPIClient';
+import { integrationTestConfig } from './integration-test.config';
+import { CoreAPIClient } from '../core-api.client';
 
 describe('DataApi', () => {
 
@@ -15,7 +15,6 @@ describe('DataApi', () => {
   }
 
   const THE_ID = CoreAPIClient.createUUID();
-
 
   it('should throw on unknown DTOs', done => {
     client.getById('THIS_IS_NOT_A_DTO' as any, 'SOME-ID')