+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] 2021-01-15
+### Added
+- batch api support 
+### Updated 
+- DTO versions, removed CrowdBusinessPartner, CrowdAssignment, CrowdPerson DTO
+### Changed
+- internal refactoring
 
 ## [1.0.1] 2020-12-11
 ### Added
-- public setter `setToken` to allow an already retrieved 
+- public `setToken` and `getToken` to allow an already retrieved token
 ### Changed
 - license to Apache-2.0